Fix vfu_dma_read/write() as per spec changes

+++ include/vfio-user.h
@@ -125,7 +125,7 @@ struct vfio_user_region_access {
 
 struct vfio_user_dma_region_access {
     uint64_t    addr;
-    uint32_t    count;
+    uint64_t    count;
     uint8_t     data[];
 } __attribute__((packed));
